Fix: add +1 to length id
