+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\index.js":"1","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\reportWebVitals.js":"2","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\App.js":"3","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\ItemList.js":"4","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\InputForm.js":"5"},{"size":496,"mtime":1610621624241,"results":"6","hashOfConfig":"7"},{"size":375,"mtime":1610621407306,"results":"8","hashOfConfig":"7"},{"size":1915,"mtime":1610625229145,"results":"9","hashOfConfig":"7"},{"size":831,"mtime":1610625243397,"results":"10","hashOfConfig":"7"},{"size":862,"mtime":1610625252744,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"5mg55y",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\index.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\reportWebVitals.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\App.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\ItemList.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\InputForm.js",[]]
+[{"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\index.js":"1","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\reportWebVitals.js":"2","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\App.js":"3","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\ItemList.js":"4","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\InputForm.js":"5","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\state\\actionTypes.js":"6","E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\state\\reducer.js":"7"},{"size":496,"mtime":1610621624241,"results":"8","hashOfConfig":"9"},{"size":375,"mtime":1610621407306,"results":"10","hashOfConfig":"9"},{"size":1369,"mtime":1610625570583,"results":"11","hashOfConfig":"9"},{"size":831,"mtime":1610625243397,"results":"12","hashOfConfig":"9"},{"size":862,"mtime":1610625252744,"results":"13","hashOfConfig":"9"},{"size":163,"mtime":1610625503775,"results":"14","hashOfConfig":"9"},{"size":529,"mtime":1610625535861,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"5mg55y",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\index.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\reportWebVitals.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\App.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\ItemList.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\components\\InputForm.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\state\\actionTypes.js",[],"E:\\Users\\Lukmaan\\Documents\\react coding challenge\\coding-challenge-react-todo-list\\src\\state\\reducer.js",[]]
