+++ src/state/actionTypes.js
@@ -1,0 +1,7 @@
+const actionTypes = {
+  ADD_ITEM: 'ADD_ITEM',
+  REMOVE_ITEM: 'REMOVE_ITEM',
+  TOGGLE_COMPLETE_ITEM: 'TOGGLE_COMPLETE_ITEM',
+};
+
+export default actionTypes;
